feat: ✨ (tools/guild) 選択している`Guild`表示コンポーネントの作成
