+++ src/ui/guilds/components/toolbar/guild/index.ts
@@ -1,0 +1,1 @@
+export { SelectedGuild } from "./guild";
