--- src/ui/guilds/components/toolbar/selectedGuild/index.ts
@@ -1,1 +1,0 @@
-export { SelectedGuild } from "./selectedGuild";
